input_heatmap

input_heatmap() takes a tbl object and easily produces a ComplexHeatmap plot, with integration with tibble and dplyr frameworks.

input_heatmap(
  .data,
  .horizontal,
  .vertical,
  .abundance,
  transform = NULL,
  .scale = "row",
  palette_abundance = c("#440154FF", "#21908CFF", "#fefada"),
  palette_grouping = list(),
  ...
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.horizontal
The name of the column horizontally presented in the heatmap

.vertical
The name of the column vertically presented in the heatmap

.abundance
The name of the transcript/gene abundance column

transform
A function, used to transform .value, for example log1p

.scale
A character string. Possible values are c(\"none\", \"row\", \"column\", \"both\")

palette_abundance
A character vector, or a function for higher customisation (colorRamp2). This is the palette that will be used as gradient for abundance. If palette_abundance is a vector of hexadecimal colours, it should have 3 values. If you want more customisation, you can pass to palette_abundance a function, that is derived as for example `colorRamp2(c(-2, 0, 2), palette_abundance)`

palette_grouping
A list of character vectors. This is the list of palettes that will be used for grouping

...
Further arguments to be passed to ComplexHeatmap::Heatmap

Value

A `ComplexHeatmap` object

Details

To be added.